requirements.txt
Django==2.1.*
djangorestframework==3.9.*
drf-nested-routers==0.91.*
Pillow==6.2.*
requests==2.22.*
